requirements.txt
./wheels/ndi_python-5.1.1.5-cp311-cp311-linux_x86_64.whl
blinker==1.7.0
Bootstrap-Flask==2.3.3
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
obsws-python==1.7.0
PyYAML==6.0.1
visitor==0.1.3
websocket-client==1.7.0
Werkzeug==3.0.1
WTForms==3.1.2
